refactor(frontend): Add EditorService
