+++ frontend/src/app/editor/model/panel.ts
@@ -1,0 +1,7 @@
+export interface Panel {
+  closed?: boolean;
+  x: number;
+  y: number;
+  rows: number;
+  cols: number;
+}
